refactor: processor_* shall be exposed
